requirements.txt
mygeotab==0.8.8
python-dotenv==1.0.0
requests==2.28.2
